Refactor Scripts Generator using factory method
