+++ SkuaScriptsGenerator/Factories/ISkuaScriptFactory.cs
@@ -1,0 +1,9 @@
+﻿using SkuaScriptsGenerator.Generators;
+
+namespace SkuaScriptsGenerator.Factories
+{
+    public interface ISkuaScriptFactory
+    {
+        ISkuaScriptWriter Generate(string type);
+    }
+}
