+++ SkuaScriptsGenerator/Factories/Writers/ISkuaScriptWriter.cs
@@ -1,0 +1,7 @@
+﻿namespace SkuaScriptsGenerator.Generators
+{
+    public interface ISkuaScriptWriter
+    {
+        void Write();
+    }
+}
